+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KV9ZK8NEHDEP09K7PR08W1YV"
+        "01KVA81EY5QB95CN85HWV2JVNK"
     )
     r"""The unique identifier of the data source"""