+++ packages/orq-rc/RELEASES.md
@@ -8769,3 +8769,13 @@ Based on:
 - [python v4.11.0-rc.26] packages/orq-rc
 ### Releases
 - [PyPI v4.11.0-rc.26] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.26 - packages/orq-rc
+
+## 2026-06-17 07:36:32
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.781.0 (2.907.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.11.0-rc.27] packages/orq-rc
+### Releases
+- [PyPI v4.11.0-rc.27] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.27 - packages/orq-rc